Teesh: Environment setup
`teesh` is responsible for setting up the bridge network which a prospect exam container and proctor container will reside within. It's also responsible for setting up the keys for SSH sessions between the two and between the host and the exam container, to allow the prospect to login.

This changes puts the management of these features into `teesh` control. `teesh` will attempt to remove any existing prospect container the setup the environment. At the successful conclusion of an exam, it will remove the exam and proctor containers.