LDAP Connector
Includes `setup_development_environment.yml` Ansible playbook to setup a
development LDAP server on a dedicated bridge network on localhost.

```
ansible-playbook setup_devlopment_environment.yml
```

This will create a small set of users, `prospect1` to `prospect3`, in
the `ou=prospects,dc=development,dc=vetzuki,dc=com` DN for testing with.

# internal/ldap.FindUser

Given a prospectID, it should find a user and return the user and a
boolean, which is true when the user is found, and false if not or an
error.