LDAP: TLS and logins
This adds support for TLS LDAP on the POC, and higher, environments. The LDAP entry seeds also have been given a password and membership into a `docker` group. The `docker` group, `cn=docker,ou=groups,dc=env,dc...`, has `memberUid` attributes which are the same as `prospect.UID`.

A visit needs to be made to the API for the `ldap.CreateUser` to add the user to the `docker` group to fully appreciate the behavior. As it stands now, the seed users can participate in the full experienc.