APIGateway config + db CNAME
This adds the proper integration method to the APIGateway instance. It attaches the `AWSLambdaBasicExecutionRole` to the `getProspect` lambda so that it can log to cloudwatch.

The integration from APIGateway is a Lambda Integration so the `getProspect` lambda takes up the whole `APIGatewayLambdaRequest` event. The event is now read for the `prospectURLID` path parameter and creates the proper errors out of `APIGatewayLambdaResponse`.

Finally, this creates a `CNAME` to the RDS database host as `db.$vetzukiEnvironment.vetuki.com`.