www.poc.vetzuki.com
This creates a custom domain mapping of `www.poc.vetzuki.com` to the ApiGateway stage hosting it associating our wildcard certificate with it.

This is mounted at the root of the api, `/`, so `curl https://www.poc.vetzuki.com` will return a missing token error as it has no methods associated with it at this time.