Model: Extend EmployerProspect
This adds `Prospect` and `Employer` property to the `EmployerProspect` type. It makes sense to serialize the `Prospect` and `Employer` along with the more abstract IDs of the `EmployerProspect`. This data isn't persisted to the database as a single row, it's the combination of all three database tables: `employer_prospect` `prospect`, and `employer`.