EmployerLogin: Valid AccessToken required
When given a valid Auth0 opaque Access Token, then a `POST /api/login` bearing it in the `Authorization` header will return the JSON serialization of an `Employer` model.

### Housekeeping

* Ignore lambda build products
* Ignore `.vscode` files

## CreateProspect: Add createProspect Lambda

`createProspect` expects an email address, a name, and a role for a given `employerID` to create a new `EmployerProspect`. The function returns JSON of an `EmployerProspect` model type.

## Deployment: Add employerLogin Lambda and route

This adds the API Gateway `POST` to `/api/login` which includes an opaque Auth0 access token as the value of the \`Authorization\` header.

## Schema

* Add `email` to `billing` table