API: Admin whitelisting
Support the environment variable `ADMIN_WHITELIST` as a command-delimited list of emails which are authorized to use the API. This is to deal with APIGateways somewhat indicipherable means of restricting access with resources policies.

This fixes a bug in the API Gateway terraform which would deploy the stage without waiting for all resources to be drawn.

The UI gets the feature of the active, completed, and total screenings scheduled to prospects.