UpdateExam: Set state by ProspectURL
This allows a `PUT` request to `/api/exams` with a `ExamStateUpdate` request to set the exam state of a Prospect

Now, when all of the docker neworks and containers are set up on the vetzuki-host, a PUT request the `/api/exams/$prospectURLID` and a body of `{"screeningState": 0|1|2|3}` will update the screening state of the Prospect.

## API Changes

The LDAP api gets \`IsGroupMember\` to check if a prospect is a member of a group.

The Model api gets \`SetPassword\` as a feature of Prospects to allow their password to be reset to a new random password.

A package \`events\` was added to the API. This allows Lambadas to use a common form of events.

## Infrastrucure

New routes and features to support /api/exams are in api_gateway and lambdas terraform configuration.

## End to End Test Case

When a prospect connects to the screening host, and all of the container features have been started, then they should be marked with their screening_state set to active (3). Then, when the prospect container exit, their screening_state should be immediately set to completed (4).