UpdateExamState: Send email on complete
When the screening_state of a prospect is set to completed, an email is sent to the prospect to let them know Vetzuki has marked it and communicated to their prospective employer.

In the same action, an email is sent to the employer with the prospect's score on the screening exam.