POC: Public launch of POC
This launches a public-facing POC using ApiGateway and S3 for static assets.

There are a few bugs, certainly, and some console messages which are out of place, also, the UI is pretty horried. Yet, the product is launched, and fixes will follow.

One notable miss is AWS simple email service is not configured