DevServer: Local development server
Reorganized each lambda to be a main binary along with a handler package. A dev server was created which wraps the lambda behavior to allow Gorilla.mux to serve the same thing locally.

Created a basic environment definition file and seeded the employer, admin@localhost, to the local database. The normal auth0 flow happens yet `admin@localhost` is always the authorized user.