Network and Port assignment
When a prospect connects to take an exam on a shared system, each prospect should have their own network for their containers and then a dedicated SSH port for their exam.

This creates a new route, `/api/networks` which accepts a `POST` request with a `prospectURLID` and `ec2InstanceID` as its payload. This creates a new network in Redis using the `ec2InstanceID` as the key and then the field `network` with an increasing counter to identify the network. The counter is split into bits and assigned from the lowest order to highest order as a network ID. In the same request, an ssh port is gained from a similar counter keyed on `ec2InstanceID` in the ssh port field. This has limitations but their unlikely to be an impact for quiet some time. A mechanism for network reclaimation could be employed later and the count of the `network` field for a given EC2 instance could be used to calculate networks available for reclaimation, or those of prospects in the `completed` state.

Teesh executes the `POST /api/networks` with an API key, pipes the output through `jq` and into a source-able shell-variables file in their configuration root. The output is sourced and integrated into their configuration. This starts to lay groundwork for the configuration being done out of the user-space and by a seperate subsystem.